Test: Alpha 0.3

+++ Server/src/lib/schema.sql
@@ -1,14 +1,13 @@
 CREATE TABLE score (
   id INT NOT NULL AUTO_INCREMENT,
   username VARCHAR(50) NOT NULL,
-  problem_0 FLOAT,
-  problem_1 FLOAT,
-  problem_3 FLOAT,
-  problem_2 FLOAT,
-  average_time FLOAT,
+  problem_0 DOUBLE(16,8),
+  problem_1 DOUBLE(16,8),
+  problem_3 DOUBLE(16,8),
+  problem_2 DOUBLE(16,8),
+  average_time DOUBLE(16,8),
   PRIMARY KEY (id)
 );
-
 UPDATE score 
 SET average_time = ( problem_1 + problem_2 + problem_3) / 3
 WHERE average_time IS NULL;
